Add new modules and tests for CIF loading, signal connections, and value conversions
- Introduced `cif_loader.pyi` for CIF file loading functionality.
- Added `signal_connection.py` to provide typed access to queued connections in PyQt6.
- Created `value_conversion.py` for numeric conversions with type checks.
- Defined `host_contract.py` to establish a typing contract for Contour widget fields.
- Implemented unit tests for CIF loading, signal connections, and value conversions.
- Added a sample CIF file for testing purposes.

+++ plugins/contour/src/contour/ui/large_dataset.py
@@ -54,4 +54,4 @@ def clamp_thumbnail_source_size(width: int, height: int) -> tuple[int, int]:
         THUMBNAIL_MAX_SOURCE_WIDTH / float(source_w),
         THUMBNAIL_MAX_SOURCE_HEIGHT / float(source_h),
     )
-    return max(1, int(round(source_w * scale))), max(1, int(round(source_h * scale)))
+    return max(1, round(source_w * scale)), max(1, round(source_h * scale))
